+++ bcbio/variation/vcfutils.py
@@ -347,6 +347,9 @@ def _sort_by_region(fnames, regions, ref_file, config):
         if fname not in added_fnames:
             if isinstance(region, (list, tuple)):
                 c, s, e = region
+            elif isinstance(region, basestring) and region.find(":") >= 0:
+                c, coords = region.split(":")
+                s, e = [int(x) for x in coords.split("-")]
             else:
                 c = region
                 s, e = 0, 0